BUGFIX
Permissions of many python files change to 600 during the apptainer build, on the server grid. Both Florian and myself are unable to replicate the issue locally, the permissions are correct when we build the apptainer image. To resolve this issue on the server I have added a command to chmod recursively and set 755 permission for each and every file in the Release directory.